refactor: 메인 레이아웃 템플릿 생성
